SmartKhata-Angular/smartkhata-ml/requirements.txt
fastapi
uvicorn
joblib
scikit-learn
pandas
psycopg2-binary